requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.18
plotly>=5.15.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
